tab panel updates and refactor
