+++ src/components/tab-panel/TabPanel.tsx
@@ -1,11 +1,12 @@
-import { Box, Typography } from "@mui/material";
+import { Box } from "@mui/material";
+
 interface TabPanelProps {
   children?: React.ReactNode;
   index: number;
   value: number;
 }
 
-export function TabPanel(props: TabPanelProps) {
+function TabPanel(props: TabPanelProps) {
   const { children, value, index, ...other } = props;
 
   return (
@@ -20,3 +21,5 @@ export function TabPanel(props: TabPanelProps) {
     </div>
   );
 }
+
+export default TabPanel;
