+++ src/components/tab-panel/index.tsx
@@ -1,0 +1,2 @@
+import TabPanel from "./TabPanel";
+export { TabPanel };
